Hardswish Activation Function added to Stateful API as it is a commonly used activation function.

+++ ivy_tests/test_ivy/test_stateful/test_activations.py
@@ -522,7 +522,7 @@ def test_relu6(
 @handle_method(
     method_tree="stateful.activations.Hardswish.__call__",
     dtype_and_x=helpers.dtype_and_values(
-        available_dtypes=helpers.get_dtypes("float64"),
+        available_dtypes=helpers.get_dtypes("float"),
         large_abs_safety_factor=8,
         small_abs_safety_factor=8,
         safety_factor_scale="log",
